+++ packages/starlight-links-validator/tests/fixtures/base-path/astro.config.ts
@@ -6,9 +6,14 @@ export default defineConfig({
   base: '/test',
   integrations: [
     starlight({
+      locales: {
+        root: { label: 'English', lang: 'en' },
+        fr: { label: 'Français', lang: 'fr' },
+      },
       pagefind: false,
-      plugins: [starlightLinksValidator()],
-      title: 'Starlight Links Validator Tests - trailing always',
+      plugins: [starlightLinksValidator({ errorOnFallbackPages: false, sameSitePolicy: 'validate' })],
+      title: 'Starlight Links Validator Tests - base path',
     }),
   ],
+  site: 'https://example.com',
 })